utilx: Move GetHostname from ginx to utilx
This introduces a new package called `utilx` which is meant for
general-purpose small utilities that others can leverage.

The first function introduced to this package is `GetHostname` which
does as its name says. However, it also added the small addition of
leveraging a `NODE_NAME` environment variable which takes precedence and
can be used to set the hostname wherever it's needed. The intended
use-case is for environments where the UTS namespace is not shared and
one cannot easily get a relevant hostname (e.g. a container).

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>