ginx: Make the routes registry helper public
The interface to be implemented to register routes to the ginx common
server used to be private. This makes it public and renames it to
`Router`. The advantage of having it public is that it makes it easier
to navigate when browsing code and viewing the documentation.

Implementations using the server should not be affected by this change.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>