correct echo group for added handlers
When a handlers Routes() method is called, `engine.Group("/")` was
called which created a new group with a prefix of `/`.

When the routes then attempted to add subgroups or routes with a prefix
of `/` such as `/user`. echo would register the route with double
slashes, resulting in an issue routing to that endpoint.

By dropping the `/` from `engine.Group("/")` it removes this duplicate
slash and resolves the routing issues.

Signed-off-by: Mike Mason <mimason@equinix.com>