Update License to Apache2.0 (#110)
The license used by Infratographer was changed from Apache 2.0 to the
Elastic License v2.0 as a method to help protect the Infratographer
project from non-intended use cases. With this change we have heard
feedback from the community that this change isn't something that is
desired. Based on this feedback we have decided to return to the
original Apache 2.0 license.

Signed-off-by: GitHub <noreply@github.com>