Add tracing to events package, deprecate TraceID/SpanID
This commit adds tracing using a new field, TraceContext, in
ChangeMessage and EventMessage. While we could stick with TraceID and
SpanID (and those fields still exist in the message and will be
marshaled/unmarshaled), there are other OpenTelemetry features like
the Baggage API and W3C traceparent/tracestate values that we can
leverage using TraceContext instead. Moving to this allows for a
tighter integration with OpenTelemetry's Go SDK as well and makes it
easier for other services using the events package to add
observability to their event systems.

Signed-off-by: John Schaeffer <jschaeffer@equinix.com>