rewrite events handling to remove watermill
As discussed in the community call, watermill doesn't give us the
necessary features we'd like to utilize with the underlying nats message.
We decided to switch to using nats directly but still wanted some
support for possibly changing this out later.

This rewrites events to use our own interfaces to allow for the
possibility of a different event driver later.

Additionally this switches to using pull subscriptions instead of push,
supports Ack, Nak and Term as well as Request/Reply semantics.

Due to the Request/Reply semantics, no longer are there separate
Publisher and Subscriber configurations as the driver needs to be able
to handle both.

Signed-off-by: Mike Mason <mimason@equinix.com>