nats: change default delivery policy (#132)
Previously the default was to set the Delivery Policy to all, however if
connecting to a consumer which already existed with a different delivery
policy, nats returns an error for a mismatch.

This change removes the default when no delivery policy is defined, to
not specify any delivery policy.

Additionally, I've added the remaining delivery policy options.

Signed-off-by: Mike Mason <mimason@equinix.com>