use nats source as connection label (#135)
Nats supports providing a connection a name.
To help with troubleshooting we should set this name so it's clear which
connections are from which service.

We already have the Source field which defines the source in the
messages, so we can simply reuse this variable and add a conneciton
option to set the name.

Signed-off-by: Mike Mason <mimason@equinix.com>