Add commonly needed message getters (#179)
What we've seen is a common usage pattern of having to build generic
functions around the usage of Change and Event messages. By adding some
commonly used functions to access data from the message, it facilitates
a better user experience when building a downstream interface to take
advantage of these fields.

---------

Signed-off-by: Tyler Auerbeck <tylerauerbeck@users.noreply.github.com>
Co-authored-by: Tyler Auerbeck <tylerauerbeck@users.noreply.github.com>