[events] Support ephemeral consumers (#184)
[loadbalancer-manager-haproxy](https://github.com/infratographer/loadbalancer-manager-haproxy)
does not have a need for durable message queues, as we want each
instance to consume all messages received related to loadbalancers.

Currently, we generate a unique durable queue for each manager instance,
which eventually puts strain on nats with the number of consumers that
can be generated over time. This PR aims to put support for ephemeral
consumers back into event subscriptions, if requested, by leaving
`QueueGroup` empty in the nats config.

---------

Signed-off-by: Matt Siwiec <rizzza@users.noreply.github.com>