Add annotation to tag an ent field as an additionalSubject (#185)
Prior to the events re-work, we were able to tag an annotation on an ent
schema field to be included as an `additionalSubject` in the event
message.

Proposing to bring that back, so any api in the infratographer ecosystem
does not have to manually manage additional subjects in the generated
event hooks.

Signed-off-by: Matt Siwiec <rizzza@users.noreply.github.com>