Allow disabling stack traces, add test.  (#192)
This adds the optional ability to turn off stack traces. It preserves
current behavior of including them.

This option is available to users that are instantiating a logger
themselves, but its not currently exposed to users of InitLoggerX

An alternative would have been to not do `lgrCfg :=
zap.NewProductionConfig()` inside the function and instead take a zap
config, but thats not the way the original code was written.

---------

Signed-off-by: Tommy Carpenter <9419125+tommyjcarpenter@users.noreply.github.com>