update code to work with keyfunc/v3
keyfunc/v3 had some significant changes which moved much of the configuration and logic to the jwkset library.
As such much of the configuration moved with it and so our configuration needed to change.

This updates the code to use this new configuration method.
Tests, with only small changes continue to pass.

Signed-off-by: Mike Mason <mimason@equinix.com>