fix goosex migration pgx not imported / update to crdbpgxv5 and pgx/v5   (#217)
This fixes pgx/v4 not being imported and upgrades to pgx/v5.

Additionally, tests have been added to ensure migrations should always
work in the future.

Finally, this also fixes issue #154 with a panic happening when goosex
which results in the logger being called but the logger not previously
being set and therefore is nil. Logger is now the Nop logger by default
to ensure no panics occur.

---------

Signed-off-by: Mike Mason <mimason@equinix.com>