fix error marshalling issues
Errors from AuthRelationshipResponse were being marshalled but due to the error interface could not be unmarshalled.
This changes the type to be a new Errors type which can handle encoding and decoding errors without resulting in unmarshalling errors.

The errors returned implement the error interface however they cannot be directly compared due to the errors being dynamically generated.

Signed-off-by: Mike Mason <mimason@equinix.com>