update echox tests to work with normalized paths
An update in v4.12.0 introduced normalized paths which ensures
empty paths or paths without a prefixed `/` start with a `/`.

This is something we ran into previously which is why we had these tests.

Now that these routes are handled properly, we needed to update our tests
as all paths are prefixed with `/` now.

Relevant PR: https://github.com/labstack/echo/pull/2616

Signed-off-by: Mike Mason <mimason@equinix.com>