fix(otelx): fix validation and ensure backwards compatibility with pre v0.7.0
otelx pre v0.7.0 used WithEndpoint rather than WithEndpointURL which now expects endpoints to be in the url format.
Using WithEndpointURL was improperly configuring the endpoint when only a hostname and port were provided.

Additionally the validation that was used prior to v0.7.0 and is still used does not properly validate the endpoint.

This change continues using WithEndpointURL but will now correct the endpoint if no scheme is provided.
This allows for backwards compatibility as well as supporting full urls.

Additionally by using this method, we're able to properly use the url.Parse to validate the url.

Signed-off-by: Mike Mason <mimason@equinix.com>