feat: add nodes field to relay connections (#471)
This adds a `nodes` field to relay connections. With this added it
allows GraphQL queries like this:

```
  location(id: $id) {
    id
    name
    devices {
      nodes {
        id
        name
      }
    }
  }
```

Which makes it easier to loop through results instead of having to loop
through the `edges` and then `node` to get the actual list of nodes.
This is similar to how GitHub and other public GraphQL APIs provide
things through the relay connection.

This is entirely based off of ent/contrib#602 and can be removed once we
are on a release that includes that PR.

Signed-off-by: Nicole Hubbard <code@nicole.dev>