fix: nodes field returning nil on nested collections (#484)
The upstream PR ent/contrib#602 appears to have missed adding the check
for the `nodes` field being used in the collections template. This is
used when you call a collection from another collection, for example:

```
query {
  loadBalancer(id: "lb-123"){
    ports {
      nodes {
        pools {
          nodes {
            id
          }
        }
      }
    }
  }
}
```

Before this fix the `pools.nodes` would come back as an empty list, with
this fix it's properly filled out.

Signed-off-by: Nicole Hubbard <code@nicole.dev>